Move getHtmlContent() to separate content script file.
